+++ crc.h
@@ -16,7 +16,7 @@ uint16_t crc_calculate (const void *buffer, size_t size);
 
 uint16_t crc (uint8_t inByte, uint16_t crc_Dbyte);
 
-uint16_t crc_bytes (const uint8_t *inBytes, unsigned int length, uint16_t crc_Dbyte);
+uint16_t crc_bytes (const uint8_t *inBytes, size_t length, uint16_t crc_Dbyte);
 
 uint16_t crc_string (const char *inString, uint16_t crc_Dbyte);
